updated CF template
